Refactor; add initial pathing logic
